Moved each component into its own file, and moved all Component files into Components directory within source and headers
