+++ headers/Components/Renderable.h
@@ -1,0 +1,25 @@
+#ifndef RENDERABLE_H
+#define RENDERABLE_H
+
+#include "Component.h"
+
+struct Renderable:Component {
+	GLuint VBO;
+	GLuint IBO;
+	int numVertices;
+	int program; // index into shadermanager
+	Material material;
+	Renderable(GLuint VBO_in, GLuint IBO_in, int vertex_Count, int program_in, Material & material_in, int ownerID):Component(ownerID)
+	{
+		VBO = VBO_in;
+		IBO = IBO_in;
+		numVertices = vertex_Count;
+		program = program_in;
+		material = material_in;
+	}
+	Renderable(){}
+
+	void readFromXML(XmlElement * element);
+};
+
+#endif
