+++ source/GameEngine/Components/Light.cpp
@@ -1,0 +1,5 @@
+#include "Components/Light.h"
+
+void Light::readFromXML(XmlElement * element) {
+	
+}
