+++ source/GameEngine/Components/ParticleSystem.cpp
@@ -1,0 +1,5 @@
+#include "Components/ParticleSystem.h"
+
+void ParticleSystem::readFromXML(XmlElement * element) {
+	
+}
