+++ source/GameEngine/Components/Player.cpp
@@ -1,0 +1,5 @@
+#include "Components/Player.h"
+
+void Player::readFromXML(XmlElement * element) {
+	
+}
