+++ source/GameEngine/Components/Renderable.cpp
@@ -1,0 +1,5 @@
+#include "Components/Renderable.h"
+
+void Renderable::readFromXML(XmlElement * element) {
+	
+}
