+++ source/GameEngine/Components/SkinnedRenderable.cpp
@@ -1,0 +1,5 @@
+#include "Components/SkinnedRenderable.h"
+
+void SkinnedRenderable::readFromXML(XmlElement * element) {
+	
+}
