+++ source/GameEngine/Components/Transform.cpp
@@ -1,0 +1,5 @@
+#include "Components/Transform.h"
+
+void Transform::readFromXML(XmlElement * element) {
+	
+}
